+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+- the plugin will now remember the SSH connection state for each workspace, and it will try to automatically
+  establish it after an expired token was refreshed.
+
 ### Fixed
 
 - `Stop` action is now available for running workspaces that have an out of date template.
 - outdated and stopped workspaces are now updated and started when handling URI
+- show errors when the Toolbox is visible again after being minimized.
 
 ## 0.3.0 - 2025-06-10